+++ src/test/java/ar/edu/grupoi/backend/desappbackend/app/DesappBackendApplicationTests.java
@@ -1,4 +1,4 @@
-package ar.edu.grupoi.backend.DesappBackend;
+package ar.edu.grupoi.backend.desappbackend.app;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
