+++ src/test/java/ar/edu/grupoi/backend/desappbackend/model/user/DonorTest.java
@@ -1,4 +1,4 @@
-package ar.edu.grupoi.backend.DesappBackend.model.user;
+package ar.edu.grupoi.backend.desappbackend.model.user;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.Mockito.mock;
@@ -7,8 +7,8 @@ import static org.mockito.Mockito.when;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import ar.edu.grupoi.backend.DesappBackend.model.project.Location;
-import ar.edu.grupoi.backend.DesappBackend.model.project.Project;
+import ar.edu.grupoi.backend.desappbackend.model.project.Location;
+import ar.edu.grupoi.backend.desappbackend.model.project.Project;
 
 
 public class DonorTest {
